+++ .env
@@ -1,4 +1,4 @@
-REACT_APP_VERSION = v1.1.2
+REACT_APP_VERSION = v0.1.0
 GENERATE_SOURCEMAP = false
 
 REACT_APP_FIREBASE_APIKEY = AIzaSyBVmQVYmHyVqLpS5UEFoPGxLPlDK8n0P_g
