+++ src/layout/MainLayout/Drawer/DrawerHeader/index.js
@@ -23,9 +23,7 @@ const DrawerHeader = ({ open }) => {
           size="small"
           sx={{ height: 16, '& .MuiChip-label': { fontSize: '0.625rem', py: 0.25 } }}
           component="a"
-          href="https://github.com/codedthemes/mantis-free-react-admin-template"
           target="_blank"
-          clickable
         />
       </Stack>
     </DrawerHeaderStyled>
