+++ src/layout/MainLayout/Header/HeaderContent/MobileSection.js
@@ -5,7 +5,6 @@ import { useTheme } from '@mui/material/styles';
 import { AppBar, Box, ClickAwayListener, IconButton, Paper, Popper, Toolbar } from '@mui/material';
 
 // project import
-import Search from './Search';
 import Profile from './Profile';
 import Transitions from 'components/@extended/Transitions';
 
@@ -86,7 +85,6 @@ const MobileSection = () => {
               <ClickAwayListener onClickAway={handleClose}>
                 <AppBar color="inherit">
                   <Toolbar>
-                    <Search />
                     <Profile />
                   </Toolbar>
                 </AppBar>
