+++ src/layout/MainLayout/index.js
@@ -47,7 +47,7 @@ const MainLayout = () => {
     <Box sx={{ display: 'flex', width: '100%' }}>
       <Header open={open} handleDrawerToggle={handleDrawerToggle} />
       <Drawer open={open} handleDrawerToggle={handleDrawerToggle} />
-      <Box component="main" sx={{ width: '100%', flexGrow: 1, p: { xs: 2, sm: 3 } }}>
+      <Box component="main" sx={{ width: '100%', flexGrow: 1, backgroundColor: theme.palette.common.white }}>
         <Toolbar />
         <Outlet />
       </Box>
