--- src/pages/extra-pages/SamplePage.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import SelectFormView from 'components/SelectFormView/index';
-
-// ==============================|| SAMPLE PAGE ||============================== //
-
-const SamplePage = () => {
-  return <SelectFormView />;
-};
-
-export default SamplePage;
