+++ src/pages/graphs/SalesColumnChart.js
