+++ src/themes/palette.js
@@ -51,7 +51,7 @@ const Palette = (mode) => {
       divider: paletteColor.grey[200],
       background: {
         paper: paletteColor.grey[0],
-        default: paletteColor.grey.A50
+        default: paletteColor.grey[0]
       }
     }
   });
